+++ tests/Feature/Console/DebugModule/Fixtures/CheckoutModule/CheckoutModuleConfig.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\DebugModule\Fixtures\CheckoutModule;
+
+use Gacela\Framework\AbstractConfig;
+
+final class CheckoutModuleConfig extends AbstractConfig
+{
+}