+++ tests/Feature/Console/DebugModule/Fixtures/CheckoutModule/CheckoutModuleFacade.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\DebugModule\Fixtures\CheckoutModule;
+
+use Gacela\Framework\AbstractFacade;
+
+/**
+ * @extends AbstractFacade<CheckoutModuleFactory>
+ */
+final class CheckoutModuleFacade extends AbstractFacade
+{
+}