+++ tests/Feature/Console/DebugModule/Fixtures/CheckoutModule/CheckoutModuleFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\DebugModule\Fixtures\CheckoutModule;
+
+use Gacela\Framework\AbstractFactory;
+
+final class CheckoutModuleFactory extends AbstractFactory
+{
+}